ref(seer): Move AutofixChatProvider to the Seer Explorer page
The provider that lets the `autofixRef` embed post a message into the chat lived inside the embed's own module and was mounted by `AssistantBlock`, one instance per assistant block, fed by a `sendMessage` prop threaded through `BlockComponent` and `BlockVariant`.

Pull the context, provider, and a `useAutofixChat` hook out into `sentry/components/seer/autofixChatContext`, and mount the provider once around the whole `SeerExplorerContent` tree. Anything rendered on the page - an embed, a block widget, the header or input section - can now call `useAutofixChat().sendMessage` to drive the agent through the chat, instead of only the embeds that happen to render inside an assistant block. The prop threading goes away with it.

`sendMessage` is still undefined on read-only runs, so entry points keep rendering disabled there, and the context value is memoized because the page re-renders on every poll.
