+++ static/app/views/seerExplorer/components/chat/assistant.tsx
@@ -4,7 +4,6 @@ import {css} from '@emotion/react';
 import {AssistantActions, AssistantMessage, MessageRow} from '@sentry/scraps/chat';
 
 import {SeerMarkdown} from 'sentry/components/seer/markdown';
-import {AutofixChatProvider} from 'sentry/components/seer/markdown/embeds/components/autofix';
 import {trackAnalytics} from 'sentry/utils/analytics';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {useSessionStorage} from 'sentry/utils/useSessionStorage';
@@ -21,7 +20,6 @@ export function AssistantBlock({
   runId,
   interactionPending,
   readOnly,
-  sendMessage,
 }: AssistantBlockProps) {
   const organization = useOrganization();
   const content = block.message.content ?? '';
@@ -41,24 +39,22 @@ export function AssistantBlock({
   }
 
   return (
-    <AutofixChatProvider sendMessage={sendMessage}>
-      <Fragment>
-        {hasValidContent(content) && (
-          <MessageRow from="assistant">
-            <AssistantMessage>
-              <SeerMarkdown raw={content} />
-            </AssistantMessage>
-          </MessageRow>
-        )}
-        <BlockActionBar
-          block={block}
-          blockIndex={blockIndex}
-          runId={runId}
-          interactionPending={interactionPending}
-          readOnly={readOnly}
-        />
-      </Fragment>
-    </AutofixChatProvider>
+    <Fragment>
+      {hasValidContent(content) && (
+        <MessageRow from="assistant">
+          <AssistantMessage>
+            <SeerMarkdown raw={content} />
+          </AssistantMessage>
+        </MessageRow>
+      )}
+      <BlockActionBar
+        block={block}
+        blockIndex={blockIndex}
+        runId={runId}
+        interactionPending={interactionPending}
+        readOnly={readOnly}
+      />
+    </Fragment>
   );
 }
